+++ src/transport/mod.rs
@@ -1,3 +1,4 @@
+pub mod dns;
 pub mod h1;
 pub mod h2;
 pub mod h3;